mqbc::WatchdogContext [new]
